Updated code to work with python server Flask.
Changed most of the HTML and CSS code to reflect the needs for the python server Flask. This includes HTML templates which other html files can inherit from.

Also added a batch script to start the server and load the hompage.html file for users who will not know how to start a flask server. NOTE: I would also like to make the batch script download Python for the correct os and pip install Flask if it's not already on the users computer.